Rename wallet module file for clarity and future refactoring.
The wallet module file has been renamed to `wallet_old.rs` to improve clarity and allow for future restructuring or updates. This change supports better organization and potential replacement with a new implementation.
