+++ src/wallet/mod.rs
@@ -1,1 +1,1 @@
-mod wallet;
+mod wallet_old;
